added a separate file for variational util stuff
